added bobomb patroling action

+++ src/game/ObjectHelpers.js
@@ -131,7 +131,7 @@ export const cur_obj_reflect_move_angle_off_wall = () => {
 }
 
 export const approach_symmetric = (value, target, increment) => {
-    const dist = target - value
+    const dist = int16(target - value)
 
     if (dist >= 0) {
         if (dist > increment) {
